moved ROS-related functionality to separate class
