+++ open-codegen/opengen/builder/__init__.py
@@ -1,3 +1,5 @@
 from .optimizer_builder import *
 from .problem import *
 from .set_y_calculator import *
+from .ros_builder import *
+
