+++ open-codegen/opengen/builder/optimizer_builder.py
@@ -11,6 +11,8 @@ import os
 import jinja2
 import logging
 
+from .ros_builder import RosBuilder
+
 _AUTOGEN_COST_FNAME = 'auto_casadi_cost.c'
 _AUTOGEN_GRAD_FNAME = 'auto_casadi_grad.c'
 _AUTOGEN_PNLT_CONSTRAINTS_FNAME = 'auto_casadi_mapping_f2.c'
@@ -527,105 +529,15 @@ class OpEnOptimizerBuilder:
         with open(target_scr_lib_rs_path, "w") as fh:
             fh.write(output_template)
 
-    def __generate_ros_dir_structure(self):
-        self.__logger.info("(ROS) Generating directory structure")
+    def __generate_ros_launch_file(self):
+        self.__logger.info("Generating open_optimizer.launch")
         target_ros_dir = self.__ros_target_dir()
-        make_dir_if_not_exists(target_ros_dir)
-        make_dir_if_not_exists(os.path.abspath(os.path.join(target_ros_dir, 'include')))
-        make_dir_if_not_exists(os.path.abspath(os.path.join(target_ros_dir, 'extern_lib')))
-        make_dir_if_not_exists(os.path.abspath(os.path.join(target_ros_dir, 'src')))
-        make_dir_if_not_exists(os.path.abspath(os.path.join(target_ros_dir, 'msg')))
-
-    def __generate_ros_package_xml(self):
-        target_ros_dir = self.__ros_target_dir()
-        template = OpEnOptimizerBuilder.__get_template('ros/package.xml')
+        template = OpEnOptimizerBuilder.__get_template('ros/open_optimizer.launch')
         output_template = template.render(meta=self.__meta)
-        target_rospkg_path = os.path.join(target_ros_dir, "package.xml")
-        with open(target_rospkg_path, "w") as fh:
+        target_rosnode_launch_path \
+            = os.path.join(target_ros_dir, "launch", "open_optimizer.launch")
+        with open(target_rosnode_launch_path, "w") as fh:
             fh.write(output_template)
-
-    def __generate_ros_cmakelists(self):
-        target_ros_dir = self.__ros_target_dir()
-        template = OpEnOptimizerBuilder.__get_template('ros/CMakeLists.txt')
-        output_template = template.render(meta=self.__meta)
-        target_rospkg_path = os.path.join(target_ros_dir, "CMakeLists.txt")
-        with open(target_rospkg_path, "w") as fh:
-            fh.write(output_template)
-
-    def __copy__ros_files(self):
-        self.__logger.info("(ROS) Copying external dependencies")
-        # 1. --- copy header file
-        target_ros_dir = self.__ros_target_dir()
-        header_file_name = self.__meta.optimizer_name + '_bindings.hpp'
-        target_include_filename = os.path.abspath(
-            os.path.join(
-                target_ros_dir, 'include', header_file_name))
-        original_include_file = os.path.abspath(
-            os.path.join(self.__target_dir(), header_file_name))
-        shutil.copyfile(original_include_file, target_include_filename)
-
-        # 2. --- copy library file
-        lib_file_name = 'lib' + self.__meta.optimizer_name + '.a'
-        target_lib_file_name = \
-            os.path.abspath(
-                os.path.join(
-                    target_ros_dir, 'extern_lib', lib_file_name))
-        original_lib_file = os.path.abspath(
-            os.path.join(
-                self.__target_dir(),
-                'target',
-                self.__build_config.build_mode,
-                lib_file_name))
-        shutil.copyfile(original_lib_file, target_lib_file_name)
-
-        # 3. --- copy msg file OptimizationParameters.msg
-        original_params_msg = os.path.abspath(
-            os.path.join(
-                og_dfn.templates_dir(), 'ros', 'OptimizationParameters.msg'))
-        target_params_msg = \
-            os.path.abspath(
-                os.path.join(
-                    target_ros_dir, 'msg', 'OptimizationParameters.msg'))
-        shutil.copyfile(original_params_msg, target_params_msg)
-
-        # 4. --- copy msg file OptimizationResult.msg
-        original_result_msg = os.path.abspath(
-            os.path.join(
-                og_dfn.templates_dir(), 'ros', 'OptimizationResult.msg'))
-        target_result_msg = \
-            os.path.abspath(
-                os.path.join(
-                    target_ros_dir, 'msg', 'OptimizationResult.msg'))
-        shutil.copyfile(original_result_msg, target_result_msg)
-
-    def __generate_ros_node_header(self):
-        self.__logger.info("Generating open_optimizer.hpp")
-        target_ros_dir = self.__ros_target_dir()
-        template = OpEnOptimizerBuilder.__get_template('ros/open_optimizer.hpp')
-        output_template = template.render(meta=self.__meta)
-        target_rosnode_header_path \
-            = os.path.join(target_ros_dir, "include", "open_optimizer.hpp")
-        with open(target_rosnode_header_path, "w") as fh:
-            fh.write(output_template)
-
-    def __generate_ros_node_cpp(self):
-        self.__logger.info("Generating open_optimizer.cpp")
-        target_ros_dir = self.__ros_target_dir()
-        template = OpEnOptimizerBuilder.__get_template('ros/open_optimizer.cpp')
-        output_template = template.render(meta=self.__meta)
-        target_rosnode_cpp_path \
-            = os.path.join(target_ros_dir, "src", "open_optimizer.cpp")
-        with open(target_rosnode_cpp_path, "w") as fh:
-            fh.write(output_template)
-
-    def __generate_ros_package(self):
-        self.__logger.info("Generating ROS package")
-        self.__generate_ros_dir_structure()
-        self.__generate_ros_package_xml()
-        self.__generate_ros_cmakelists()
-        self.__copy__ros_files()
-        self.__generate_ros_node_header()
-        self.__generate_ros_node_cpp()
 
     def build(self):
         """Generate code and build project
@@ -668,5 +580,5 @@ class OpEnOptimizerBuilder:
             self.__generate_c_bindings_makefile()
 
         if self.__build_config.ros_config is not None:
-            self.__generate_ros_package()
-
+            ros_builder = RosBuilder(self.__meta, self.__build_config)
+            ros_builder.build()
