+++ open-codegen/opengen/templates/ros/open_optimizer.launch
@@ -1,0 +1,10 @@
+<launch>
+  <node name="ros_node_{{meta.optimizer_name}}"
+        pkg="ros_node_{{meta.optimizer_name}}"
+        type="ros_node_{{meta.optimizer_name}}_optimizer"
+        respawn="true"
+        clear_params="true"
+        output="screen">
+    <rosparam file="$(find ros_node_{{meta.optimizer_name}})/config/open_params.yaml" />
+  </node>
+</launch>
