+++ open-codegen/opengen/templates/ros/open_params.yaml
@@ -1,0 +1,3 @@
+solution_topic: "solution"
+params_topic: "parameters"
+rate: 40
